+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 - Added support for date values formatted as "YYYYMMDD"
 - Pressing the button will now clear the current selection before navigating back
 - Added support for non-standard `ARTISTS` tags
-- Reworked music folders dialog to be more coherent
+- Reworked music folders dialog to be more visually straightforward
+- Play Next and Add To Queue now start playback if there is no queue to add
 
 #### What's Fixed
 - Fixed unreliable ReplayGain adjustment application in certain situations